삭제 기능 사전 검증
1. 게시글 실존 여부
2. 내가 쓴 글이 맞는지?

+++ server/api/post/post.py
@@ -80,6 +80,22 @@ def delete_post(params):
     
     # 파라미터 검증 -> 본인이 쓴 글이 맞는지? + 지우려는 글이 실존하는지?
     
+    sql = f"SELECT * FROM posts WHERE id = {params['post_id']}"
+    
+    post_data = db.executeOne(sql)
+    
+    if post_data == None:
+        return {
+            'code': 400,
+            'message': '해당 게시글은 존재하지 않습니다.'
+        }, 400
+        
+    if post_data['user_id'] != int(params['user_id']):
+        return {
+            'code': 400,
+            'message': '본인이 작성한 게시글만 삭제할 수 있습니다.'
+        }, 400
+    
     sql = f"DELETE FROM posts WHERE id = {params['post_id']}"
     
     db.executeQueryAndCommit(sql)
